DOM
Dicee is a simple game where players roll two virtual dice on a website and try to match the numbers shown on the dice. It can be built using HTML, CSS, JavaScript, and the Document Object Model (DOM).

HTML is used to structure the website and create the layout for the game. It is used to create the container for the dice images and the button to roll the dice.

CSS is used to style the website and make it visually appealing. It can be used to change the background color, font size, and other visual elements of the website.

JavaScript is used to create the interactive functionality of the game. It can be used to create the code that randomizes the dice rolls and updates the images on the website to match the new numbers. JavaScript is also used to create the event listener that listens for the button click and triggers the dice roll.

The Document Object Model (DOM) is used to create a connection between the HTML and CSS elements on the website and the JavaScript code. It allows the JavaScript code to access and manipulate the elements on the website, such as changing the images of the dice when the roll button is clicked.

Overall, Dicee game is a simple game that can be built using HTML, CSS, JavaScript, and the DOM, where the HTML and CSS are used for the layout and styling of the website, JavaScript for the interactive functionality and the DOM is used to connect the HTML, CSS and javascript together.